Rename modules/http_api/httpApi.js -> modules/http_api/http_api.js
Convention restricts file names to be written only in snake case format.
